requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# LangGraph and LangChain
langgraph>=0.0.40
langchain>=0.1.0
langchain-openai>=0.1.0

# PDF Processing
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
pdf2image>=3.1.0

# AI and ML
openai>=1.0.0
tiktoken>=0.5.0

# Data Processing
pydantic>=2.0.0
pandas>=2.0.0
numpy>=1.24.0

# Database
sqlalchemy>=2.0.0
sqlite3

# Web and Templates
jinja2>=3.1.0
aiofiles>=23.0.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
typing-extensions>=4.8.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

